+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `MLBMatchupComparison` for fox sports json response for probable starting pitchers (#73)
+- `MLBProbableStartingPitcher` fox sports data model for probable starting pitcher (#73)
+- Scraping fox sports MLB probable starting pitcher using `MLBProbableStartingPitcherScraper` (#73)
 
 ## [0.8.2] - 2025-05-17
 ### Fixed